gizmos refactor
* Changed the way Samples interact with Gizmos
* Simplified the Gizmos system
* Adjust namespaces
