+++ TGC.MonoGame.Samples/Samples/Physics/BEPU/WallOfBoxes.cs
@@ -13,7 +13,7 @@ using TGC.MonoGame.Samples.Physics.Bepu;
 using TGC.MonoGame.Samples.Viewer;
 using NumericVector3 = System.Numerics.Vector3;
 
-namespace TGC.MonoGame.Samples.Samples.Physics.Bepu
+namespace TGC.MonoGame.Samples.Samples.Physics.BEPU
 {
     /// <summary>
     ///     Wall Of Boxes:
