+++ TGC.MonoGame.Samples/Viewer/Gizmos/Geometries/CubeGizmoGeometry.cs
@@ -1,10 +1,7 @@
 ﻿using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace TGC.MonoGame.Samples.Viewer
+namespace TGC.MonoGame.Samples.Viewer.Gizmos.Geometries
 {
     /// <summary>
     ///     Gizmo for drawing Wire Cubes.
