+++ TGC.MonoGame.Samples/Viewer/Gizmos/Geometries/GizmoGeometry.cs
@@ -1,10 +1,6 @@
-﻿using Microsoft.Xna.Framework;
-using Microsoft.Xna.Framework.Graphics;
-using System;
-using System.Collections.Generic;
-using System.Text;
+﻿using Microsoft.Xna.Framework.Graphics;
 
-namespace TGC.MonoGame.Samples.Viewer
+namespace TGC.MonoGame.Samples.Viewer.Gizmos.Geometries
 {
     /// <summary>
     ///     Gizmo that is drawn using non-user indexed line lists.
