+++ TGC.MonoGame.Samples/Viewer/Gizmos/Geometries/RadialGizmoGeometry.cs
@@ -1,15 +1,13 @@
-﻿using Microsoft.Xna.Framework;
+﻿using System;
+using Microsoft.Xna.Framework;
 using Microsoft.Xna.Framework.Graphics;
-using System;
-using System.Collections.Generic;
-using System.Text;
 
-namespace TGC.MonoGame.Samples.Viewer
+namespace TGC.MonoGame.Samples.Viewer.Gizmos.Geometries
 {
     /// <summary>
     ///     Abstract class for any geometry that contains a circle.
     /// </summary>
-    abstract internal class RadialGizmoGeometry : GizmoGeometry
+    internal abstract class RadialGizmoGeometry : GizmoGeometry
     {
         /// <summary>
         ///     Creates a Radial Geometry, which contains a radial wire shape.
